fix(app): harden bash exec governance
