+++ docs/plans/2026-03-29-bash-governance-ast-prefix-rule-implementation-plan.md
@@ -143,7 +143,7 @@ fn tool_runtime_config_projects_bash_rules_dir_override() {
 
     assert_eq!(
         runtime.bash_exec.governance.rules_dir,
-        std::path::PathBuf::from("/tmp/work/custom/rules")
+        std::path::PathBuf::from("custom/rules")
     );
 }
 ```
